requirements.txt
Django==3.2.3
django-ninja==0.13.2